Add transactions
This is a pain in the arse but generally should allow us to use Sentry's
concept of web transactions for things like hooks.

It's done in a stack to support instances where one hook may call
another hook etc. Anything that doesn't override bubbles upwards.

This (unfortunately) does *NOT* support callbacks, meaning that for now
timers will be on their own. A fix for this is Coming Soon™

This also requires manual marking up of transactions, but doing them
automatically is next on the list.