Automatic transactions!
This is terrifying and might be really bad for performance in the hooks
department, but let's give it a go!

This will automatically create transactions for all the routes into Lua
that I can mess with - though timers and entity callbacks are still
mysteries.

It also doesn't handle async boundries, which is an "oh god" of a todo.